Avoid Year 2038 compatibility issues

+++ daemons/controld/controld_execd_state.c
@@ -75,8 +75,8 @@ fail_pending_op(gpointer key, gpointer value, gpointer user_data)
     event.interval_ms = op->interval_ms;
     event.rc = PCMK_OCF_UNKNOWN_ERROR;
     event.op_status = PCMK_LRM_OP_NOT_CONNECTED;
-    event.t_run = op->start_time;
-    event.t_rcchange = op->start_time;
+    event.t_run = (unsigned int) op->start_time;
+    event.t_rcchange = (unsigned int) op->start_time;
 
     event.call_id = op->call_id;
     event.remote_nodename = lrm_state->node_name;
